Levels inside the PieSlice are operatives.
It has been necessary to create a mechanism to allow notifications
towards PieSlice from Slice when it is clicked.

Details:
Slice has registered a method of PieSlice that is called when the
click event is received. Then PieSlice will know the level that has
been clicked and can turnOn/turnOff slices according to the level.